+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Only fetch testtuples when building charts (#135)
 -   Chart tooltips are now interactive (#142)
 -   All algos are now listed in the same table (#128)
+-   Display current node id on login page (#148)
 
 ## [0.2.0] - 2021-09-02